requirements.txt
# Core dependencies
python-telegram-bot==20.7
python-dotenv==1.0.0

# Web automation
selenium==4.15.2
webdriver-manager==4.0.1

# Data processing
pandas==2.1.3
openpyxl==3.1.2
lxml==4.9.3

# Optional: OCR capabilities (for CAPTCHA solving)
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0

# Utilities
requests==2.31.0
urllib3==2.1.0 